custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-type-extension-is-a-type.html

<!DOCTYPE html>
<html>
<head>
<title>If typeExtension presents in createElement call then typeExtension is a type</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If typeExtension presents in createElement call then typeExtension is a type">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        var obj = doc.createElement(tagName);
        var name = 'x-a-' + tagName;
        var proto = Object.create(obj.constructor.prototype);
        var GeneratedConstructor = doc.registerElement(name, {prototype: proto, extends: tagName});
        var customElement = doc.createElement(tagName, name);

        assert_equals(Object.getPrototypeOf(customElement), GeneratedConstructor.prototype,
            'Custom element type should be the type extension in createElement call');
    });
}, 'If typeExtension presents in createElement call then type is a type extension in createElement call');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        var obj = doc.createElement(tagName);
        var name = 'x-b-' + tagName;
        var proto = Object.create(obj.constructor.prototype);
        var GeneratedConstructor = doc.registerElement(name, {prototype: proto, extends: tagName});
        var customElement = doc.createElementNS(HTML_NAMESPACE, tagName, name);

        assert_equals(Object.getPrototypeOf(customElement), GeneratedConstructor.prototype,
            'Custom element type should be the type extension in createElement call');
    });
}, 'If typeExtension presents in createElementNS call then type is a type extension in createElement call');
</script>
</body>
</html>
